fix(window): add evaluate_expression wrapper and fix test assertions
- Add `evaluate_expression` wrapper function in window/utils.rs that
  calls `evaluate_expression_with_map` with an empty HashMap, providing
  the simpler callback signature needed by tests
- Fix FIRST_VALUE/LAST_VALUE test assertions: when sorted by salary DESC,
  higher values (90000) come first, not lower values (85000)

+++ tests/executor/window_functions.rs
@@ -481,10 +481,10 @@ fn test_window_first_value() {
         assert_eq!(result[0].values[3], SqlValue::Integer(120000));
         assert_eq!(result[1].values[3], SqlValue::Integer(120000));
 
-        // Sales rows should have top_salary = 85000 (Carol is first in DESC order)
-        // Note: Carol (85000) comes before Dave (90000) when sorted by string comparison
-        assert_eq!(result[2].values[3], SqlValue::Integer(85000));
-        assert_eq!(result[3].values[3], SqlValue::Integer(85000));
+        // Sales rows should have top_salary = 90000 (Dave is first in DESC order)
+        // When sorted by salary DESC: Dave (90000) comes before Carol (85000)
+        assert_eq!(result[2].values[3], SqlValue::Integer(90000));
+        assert_eq!(result[3].values[3], SqlValue::Integer(90000));
 
         println!("✅ FIRST_VALUE test works!");
     } else {
@@ -568,10 +568,10 @@ fn test_window_last_value() {
         assert_eq!(result[0].values[3], SqlValue::Integer(95000));
         assert_eq!(result[1].values[3], SqlValue::Integer(95000));
 
-        // Sales rows should have lowest_salary = 90000 (Dave is last in DESC order)
-        // Note: Dave (90000) comes after Carol (85000) when sorted by string comparison
-        assert_eq!(result[2].values[3], SqlValue::Integer(90000));
-        assert_eq!(result[3].values[3], SqlValue::Integer(90000));
+        // Sales rows should have lowest_salary = 85000 (Carol is last in DESC order)
+        // When sorted by salary DESC: Dave (90000) comes before Carol (85000)
+        assert_eq!(result[2].values[3], SqlValue::Integer(85000));
+        assert_eq!(result[3].values[3], SqlValue::Integer(85000));
 
         println!("✅ LAST_VALUE test works!");
     } else {
